+++ afxdp/server_main.cc
@@ -165,7 +165,7 @@ int server_init(struct server_t* server, const char* interface_name) {
     // per-CPU socket setup
     for (int i = 0; i < MY_NUM_QUEUES; i++) {
         // allocate umem_buffer for umem
-        const int buffer_size = NUM_FRAMES * FRAME_SIZE;
+        const size_t buffer_size = NUM_FRAMES * FRAME_SIZE;
 
         if (posix_memalign(&server->socket[i].umem_buffer, getpagesize(),
                            buffer_size)) {
